added variational qoi class now looking into some unit tests
